+++ SleepModel/HEMListItemSelectionViewController.m
@@ -11,7 +11,7 @@
 #import "HEMBaseController+Protected.h"
 #import "HEMActivityIndicatorView.h"
 
-@interface HEMListItemSelectionViewController()
+@interface HEMListItemSelectionViewController() <HEMListPresenterDelegate>
 
 // FIXME: this extra navigation bar is a workaround for the mess that is
 // the alarm code.  Once we rewrite the alarm code, we should consider
@@ -43,7 +43,16 @@
     [[self listPresenter] bindWithNavigationBar:[self extraNavigationBar]
                               withTopConstraint:[self navigationBarTopConstraint]];
     [[self listPresenter] bindWithActivityIndicator:[self activityIndicator]];
+    [[self listPresenter] setPresenterDelegate:self];
     [self addPresenter:[self listPresenter]];
+}
+
+#pragma mark - Presenter Delegate
+
+- (void)presentErrorWithTitle:(NSString *)title
+                      message:(NSString *)message
+                         from:(HEMListPresenter *)presenter {
+    [self showMessageDialog:message title:title];
 }
 
 @end