+++ SleepModel/HEMSleepSoundConfigurationCell.m
@@ -24,6 +24,7 @@ static CGFloat const HEMSleepSoundGraphMinScale = 0.3f;
     [super awakeFromNib];
     
     [[self titleLabel] setFont:[UIFont cardTitleFont]];
+    [[self playingLabel] setFont:[UIFont cardTitleFont]];
     
     [[self titleSeparator] setBackgroundColor:[UIColor separatorColor]];
     [[self soundSeparator] setBackgroundColor:[UIColor separatorColor]];